[SofaAssimp] Clean some warnings.

+++ applications/plugins/SofaAssimp/AssimpLoader.cpp
@@ -145,8 +145,8 @@ bool AssimpLoader::convertAssimpScene()
         aiMesh* currentMesh = m_assimpScene->mMeshes[i]; //The ith mesh of the array of meshes.
         unsigned int nbr_pos = currentMesh->mNumVertices;
         
-        unsigned int cpt_pos = waPositions.size();
-        unsigned int cpt_norm = waNormals.size();
+        unsigned int cpt_pos = (unsigned int)waPositions.size();
+        unsigned int cpt_norm = (unsigned int)waNormals.size();
         //unsigned int cpt_uv = m_texCoords.size();
 
         if (cpt_pos != cpt_norm /*|| cpt_pos != cpt_uv*/)
